feat: added the Similar Cards Component
